+++ bee-ledger/bee-ledger-types/src/rand/output_diff.rs
@@ -1,11 +1,11 @@
 // Copyright 2020-2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use bee_ledger_types::{OutputDiff, TreasuryDiff};
-
-use crate::rand::{
+use bee_block::rand::{
     milestone::rand_milestone_id, number::rand_number_range, option::rand_option, output::rand_output_id,
 };
+
+use crate::{OutputDiff, TreasuryDiff};
 
 /// Generates a random treasury diff.
 pub fn rand_treasury_diff() -> TreasuryDiff {