+++ bee-tangle/src/rand/unreferenced_block.rs
@@ -1,9 +1,9 @@
 // Copyright 2020-2021 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-use bee_tangle::unreferenced_block::UnreferencedBlock;
+use bee_block::rand::block::rand_block_id;
 
-use crate::rand::block::rand_block_id;
+use crate::unreferenced_block::UnreferencedBlock;
 
 /// Generates a random unreferenced block.
 pub fn rand_unreferenced_block() -> UnreferencedBlock {